add cumulative score option

+++ osu.Game.Tournament/Models/TournamentMatch.cs
@@ -35,14 +35,14 @@ namespace osu.Game.Tournament.Models
 
         public string? Team1Acronym;
 
-        public readonly Bindable<int?> Team1Score = new Bindable<int?>();
+        public readonly Bindable<long?> Team1Score = new Bindable<long?>();
 
         [JsonIgnore]
         public readonly Bindable<TournamentTeam?> Team2 = new Bindable<TournamentTeam?>();
 
         public string? Team2Acronym;
 
-        public readonly Bindable<int?> Team2Score = new Bindable<int?>();
+        public readonly Bindable<long?> Team2Score = new Bindable<long?>();
 
         public readonly Bindable<bool> Completed = new Bindable<bool>();
